get profile details applied to API
