+++ README.md
@@ -12,18 +12,12 @@ Building SkypeClone using Material-UI, React.js, Express.js, Node.js, MongoDB &a
 ---
 - Mir
 - Michal
-<<<<<<< HEAD
-
 ---
-## Data Temp 
- For the porpuse of the examine the Schema of the mongoose we created the `data.js` file in `temporary` folder.
 
-=======
----
 ## Data Temp
  For the porpuse of the examine the Schema of the mongoose we created the `data.js` file in `temporary` folder.
->>>>>>> ab55dc523144a86a345820f7a402c66f05e8f0af
- -To add the data to your ``Database`` apply the following lines in your mongo database:
+ -To add the data to your ``Database`` apply the following lines in your mongo database in `Terminal`:
+
   ````
    use skypeClone
    db.createCollection('users')
