+++ app.js
@@ -1,45 +1,44 @@
-var express = require('express');
-var app = express();
-var bodyParser = require('body-parser');
-var Message = require('./models/message');
-var mongo = require('mongodb');
-var mongoose = require('mongoose');
+const express = require('express');
+const app = express();
+const mongoose = require('mongoose');
+const bodyParser = require('body-parser');
+const Contact = require('./models/user.model');
+const db = 'mongodb://localhost:27017/skypeClone';
+const port = process.env.PORT || 3000;
 
-var db = 'mongodb://localhost:27017/skypeClone';
 mongoose.Promise = global.Promise;
 mongoose.connection.openUri(db);
 
 app.use(bodyParser.json());
-app.use(bodyParser.urlencoded({ extended: false }));
+app.use(bodyParser.urlencoded({
+    extended: true
+}));
 
-app.post('/message', (req, res) => {
-	var newMessage = new Message ({
-		senderUserID: req.body.sendUserId,
-		receiverUserID: req.body.receiveUserId,
-		message: req.body.textMessage
-	});
-	newMessage.save((err, message) => {
-		if(err) {
-	      res.send('Error!');
-	    } else {
-	      res.send(message);
-	    }
-	});
+app.get('/get_users/:id', (request, response)=>{
+    Contact.findOne({
+        _id: request.params.id
+    }).exec((err, contact)=>{
+        if(err){
+            response.send('We can not find the friends');
+        }else{
+             response.json(contact.contacts.list);
+        }
+    });
 });
 
-app.get('/:message', (req, res) => {
-	Message.findOne({
-		message: req.params.message
-	})
-	.exec((err, message) => {
-		if(err) {
-			res.send('Error!');
-		} else {
-			res.json(message);
-		}
-	});
+app.get('/get_profile_detail/:id', (request, response)=>{
+     Contact.findOne({
+        _id: request.params.id
+     }).exec((err, contact)=>{
+        if(err){
+            response.send('Err');
+        }else{
+            response.json(contact);
+        }
+     }) 
 });
 
-app.listen(3000, function() {
-	console.log('Server started .....')
+app.listen(port, ()=>{
+    console.log('Server is listening to port ' + port);
 });
+
