+++ models/user.model.js
@@ -29,4 +29,4 @@ const userSchema = new Schema({
 	}
 });
 
-module.exports = mongoose.model('user', userSchema);
+module.exports = mongoose.model('users', userSchema);
